2020.9.2 add club data parser

+++ app/src/main/java/com/example/brawlkat/kat_database.java
@@ -50,11 +50,17 @@ public class kat_database extends SQLiteOpenHelper {
 
         SQLiteDatabase database = getWritableDatabase();
 
-        //database.execSQL("DELETE FROM katabase WHERE searchTag = " + "'" + Tag + "'");
-
+        // 일단 전체 테이블을 선택하고 커서를 마지막으로 보낸다.
         Cursor cursor = database.rawQuery("SELECT * FROM katabase", null);
         cursor.moveToLast();
 
+        // 테이블이 비었으면 그냥 삽입하고 종료
+        if(cursor.getCount() <= 0){
+            database.execSQL("INSERT INTO katabase VALUES(null, '" + Type + "', '" + Tag + "', '" + Account + "', '" + isAccount + "');");
+            return;
+        }
+
+        // 테이블이 비어있지 않다면 현재 저장하고자 하는 데이터와 중복되는 데이터가 있는지 체크
         do{
             if(cursor.getString(2).equals(Tag)){
                 return;
@@ -62,39 +68,44 @@ public class kat_database extends SQLiteOpenHelper {
         }
         while(cursor.moveToPrevious());
 
-        database.execSQL("INSERT INTO katabase VALUES(null, '" + Type + "', '" + Tag + "', '" + Account + "', '" + isAccount + "'); " +
-                "WHERE ");
+        // 중복되는 데이터가 없으면 그냥 삽입
+        database.execSQL("INSERT INTO katabase VALUES(null, '" + Type + "', '" + Tag + "', '" + Account + "', '" + isAccount + "');");
     }
 
-    public ArrayList<ArrayList<String> > get(){
+    public ArrayList<ArrayList<String> > get(String type){
 
         ArrayList<ArrayList<String> > resultList = new ArrayList<>();
 
         SQLiteDatabase database = getWritableDatabase();
 
-        Cursor cursor = database.rawQuery("SELECT * FROM katabase", null);
+        Cursor cursor = database.rawQuery("SELECT * FROM katabase WHERE searchType = " + "'" + type + "'", null);
         cursor.moveToLast();
 
-        do{
-            ArrayList<String> inner = new ArrayList<>();
-            for(int i = 1; i < 5; i++){
-                inner.add(cursor.getString(i));
-            }
+        if(cursor.getCount() <= 0) return resultList;
+        do {
+            if(cursor.getString(1).equals(type)) {
+                ArrayList<String> inner = new ArrayList<>();
+                for (int i = 1; i < 5; i++) {
+                    inner.add(cursor.getString(i));
+                }
 
-            resultList.add(inner);
+                resultList.add(inner);
+            }
         }
-        while(cursor.moveToPrevious());
+        while (cursor.moveToPrevious());
+
 
         return resultList;
     }
 
-    public void delete(){
+    public void delete(String type){
         SQLiteDatabase database = getWritableDatabase();
 
         if(size() > 9){
             Cursor cursor = database.rawQuery("SELECT * FROM katabase", null);
             cursor.moveToFirst();
 
+            if(cursor.isFirst()) return;
             String currTag = cursor.getString(2);
             database.execSQL("DELETE FROM katabase WHERE searchTag = " + "'" + currTag + "'");
         }
@@ -109,9 +120,12 @@ public class kat_database extends SQLiteOpenHelper {
         Cursor cursor = database.rawQuery("SELECT * FROM katabase", null);
         cursor.moveToLast();
 
-        do{
-            System.out.println(cursor.getString(2));
+        if(cursor.getCount() > 0) {
+            do {
+                System.out.println(cursor.getString(1) + ", " + cursor.getString(2) + ", " + cursor.getString(3) + ", "
+                        + cursor.getString(4));
+            }
+            while (cursor.moveToPrevious());
         }
-        while(cursor.moveToPrevious());
     }
 }
